Put workFrom one abstract class higher and avoid to stand on saplings.
Also move some things to the miner we don't need in the structure
