+++ src/main/java/com/minecolonies/coremod/entity/ai/citizen/lumberjack/EntityAIWorkLumberjack.java
@@ -6,6 +6,7 @@ import com.minecolonies.coremod.entity.ai.util.AIState;
 import com.minecolonies.coremod.entity.ai.util.AITarget;
 import com.minecolonies.coremod.entity.pathfinding.PathJobFindTree;
 import com.minecolonies.coremod.util.BlockPosUtil;
+import com.minecolonies.coremod.util.MathUtils;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockSapling;
 import net.minecraft.block.SoundType;
@@ -46,6 +47,11 @@ public class EntityAIWorkLumberjack extends AbstractEntityAIInteract<JobLumberja
     private static final int SEARCH_INCREMENT = 5;
 
     /**
+     * Position where the Builders constructs from.
+     */
+    private BlockPos  workFrom;
+
+    /**
      * If this limit is reached, no trees are found.
      */
     private static final int SEARCH_LIMIT = 150;
@@ -70,6 +76,16 @@ public class EntityAIWorkLumberjack extends AbstractEntityAIInteract<JobLumberja
      * The speed in which he backs away.
      */
     private static final double WALK_BACK_SPEED = 1.0;
+
+    /**
+     * The standard range the lumberjack should reach until his target.
+     */
+    private static final int STANDARD_WORKING_RANGE = 2;
+
+    /**
+     * The minimum range the lumberjack has to reach in order to construct or clear.
+     */
+    private static final int MIN_WORKING_RANGE      = 2;
 
     /**
      * Time in ticks to wait before placing a sapling.
@@ -306,6 +322,39 @@ public class EntityAIWorkLumberjack extends AbstractEntityAIInteract<JobLumberja
     }
 
     /**
+     * Calculates the working position.
+     * <p>
+     * Takes a min distance from width and length.
+     * <p>
+     * Then finds the floor level at that distance and then check if it does contain two air levels.
+     * @param targetPosition the position to work at.
+     * @return BlockPos position to work from.
+     */
+    @Override
+    public BlockPos getWorkingPosition(final BlockPos targetPosition)
+    {
+        return getWorkingPosition(1, targetPosition, 0);
+    }
+
+    /**
+     * Walk to the current construction site.
+     * <p>
+     * Calculates and caches the position where to walk to.
+     * @param workAt block to work at.
+     * @return true while walking to the site.
+     */
+    public boolean walkToTree(final BlockPos workAt)
+    {
+        if (workFrom == null)
+        {
+            workFrom = getWorkingPosition(workAt);
+        }
+
+        //The miner shouldn't search for a save position. Just let him build from where he currently is.
+        return worker.isWorkerAtSiteWithMove(workFrom, STANDARD_WORKING_RANGE) || MathUtils.twoDimDistance(worker.getPosition(), workFrom) < MIN_WORKING_RANGE;
+    }
+
+    /**
      * Work on the tree.
      * First find your way to the tree trunk.
      * Then chop away
@@ -317,7 +366,7 @@ public class EntityAIWorkLumberjack extends AbstractEntityAIInteract<JobLumberja
     private AIState chopTree()
     {
         final BlockPos location = job.tree.getLocation();
-        if (walkToBlock(location))
+        if (walkToTree(location))
         {
             checkIfStuckOnLeaves(location);
             return getState();
@@ -331,12 +380,13 @@ public class EntityAIWorkLumberjack extends AbstractEntityAIInteract<JobLumberja
             }
             plantSapling();
             this.getOwnBuilding().getColony().incrementStatistic("trees");
+            workFrom = null;
             return LUMBERJACK_GATHERING;
         }
 
         //take first log from queue
         final BlockPos log = job.tree.peekNextLog();
-        if (!mineBlock(log))
+        if (!mineBlock(log, workFrom))
         {
             return getState();
         }
